+++ app/ExternalFiles/Extensions/Export_By_Size.php
@@ -111,7 +111,7 @@ class Export_By_Size extends Extension_Base {
 
 		// get the settings page.
 		$settings_page = $settings_obj->get_page( \ExternalFilesInMediaLibrary\Plugin\Settings::get_instance()->get_menu_slug() );
-		if( ! $settings_page instanceof Page ) {
+		if ( ! $settings_page instanceof Page ) {
 			return;
 		}