+++ changelog.md
@@ -80,6 +80,7 @@
 - Added configurations to quickly configure the plugin for a specific task
 - Added output of all (not private) settings in site health report
 - Added site health checks for cronjobs, and the availability of external sources
+- Added option to export the installation hash key, which is used for any encryption
 
 ### Changed